rename demo to MessageExchangeBus
