+++ Assets/Tests/Editor/TestJson.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
-using ClearScriptDemo.Demo.SpawnDemo;
-using ClearScriptDemo.JSonConverters;
+using ClearScriptDemo.Demo.MessageExchangeBus;
+using ClearScriptDemo.Demo.MessageExchangeBus.Utils;
 using Newtonsoft.Json;
 using NUnit.Framework;
 using UnityEngine;
