+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,6 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Demo/Scenes/SpawnDemo.unity
+    path: Assets/Demo/Scenes/MessageExchangeBusDemo.unity
     guid: d8f1e3f4c28256c4fa504267b1b8d677
   m_configObjects: {}
